Added Generic Tool and toolbar
MainScene now contains a toolbar and two GenericTool objects; Game Prefabs folder contains the GenericTool prefab. GenericTools are controlled with LMB; click on them to pick them up, click on the toolbar to put them down, click on the screen to use them (the last one has no effect so far; the method is intended for VFX and SFX).