Tool and Cursor issue
Fixed a bug where the cursor was interfering with the toolbar, making it so using a tool would put it away every time.  A workaround was in place where we simply offset the cursor from the tool, but this was sub-optimal.  This change now has the cursor in the correct place, fixed the issue with the tray, and allows the fire extinguisher to properly work.